refactor(core): extract media generation into a dedicated provider service
local-provider-service.ts had grown to 1,440 lines after absorbing media
eligibility, resolution, catalog, persistence, and execution alongside its
existing provider/voice/transcription duties. Move those concerns into
media-generation-service.ts (isUsableImageGenerationModel,
resolveConfiguredMediaGenerationTarget, buildMediaGenerationModelCatalog,
resolveActiveMediaGenerationSettings, saveMediaGenerationSettings,
clearMediaGenerationSelections, generateConfiguredMedia) and the shared
model-map resolution into provider-model-map.ts. local-provider-service.ts
is back to 1,185 lines, near its pre-feature size.

The media contract tests move to media-generation-service.test.ts and gain
direct coverage for the catalog builder, active-settings resolution, and
selection clearing. @cline/core package exports are unchanged; only the
internal module layout moved, per the SDK refactor standard (no shims).
