+++ sdk/packages/core/src/index.ts
@@ -675,24 +675,17 @@ export {
 	type DeleteLocalProviderRequest,
 	deleteLocalProvider,
 	ensureCustomProvidersLoaded,
-	type GenerateConfiguredMediaRequest,
-	type GenerateConfiguredMediaResult,
-	generateConfiguredMedia,
 	getLocalProviderModels,
 	isDedicatedTranscriptionModel,
-	isUsableImageGenerationModel,
 	listLocalProviders,
 	loginAndSaveLocalProviderOAuthCredentials,
 	loginLocalProvider,
 	markLocalProviderEnabled,
 	normalizeOAuthProvider,
-	type ResolvedMediaGenerationTarget,
 	refreshProviderModelsFromSource,
-	resolveConfiguredMediaGenerationTarget,
 	resolveLocalClineAuthToken,
 	saveLocalProviderOAuthCredentials,
 	saveLocalProviderSettings,
-	saveMediaGenerationSettings,
 	saveVoiceInputSettings,
 	type TranscribeConfiguredVoiceInputRequest,
 	type TranscribeLocalAudioRequest,
@@ -701,6 +694,16 @@ export {
 	type UpdateLocalProviderRequest,
 	updateLocalProvider,
 } from "./services/providers/local-provider-service";
+export {
+	buildMediaGenerationModelCatalog,
+	type GenerateConfiguredMediaRequest,
+	type GenerateConfiguredMediaResult,
+	generateConfiguredMedia,
+	isUsableImageGenerationModel,
+	type ResolvedMediaGenerationTarget,
+	resolveConfiguredMediaGenerationTarget,
+	saveMediaGenerationSettings,
+} from "./services/providers/media-generation-service";
 export {
 	getProviderConfigFields,
 	type ProviderConfigFieldKey,
