+++ sdk/packages/core/src/services/providers/local-provider-service.ts
@@ -1,24 +1,15 @@
 import * as LlmsModels from "@cline/llms";
-import {
-	type AddProviderActionRequest,
-	type AgentUsage,
-	type ITelemetryService,
-	MEDIA_GENERATION_TYPES,
-	type MediaContent,
-	type MediaGenerationModelCatalog,
-	type MediaGenerationSettings,
-	type MediaGenerationType,
-	type MediaModelSelection,
-	modelProducesImages,
-	type ProviderCapability,
-	type ProviderCatalogResponse,
-	type ProviderConfigField,
-	type ProviderConfigFieldPrimitive,
-	type ProviderListItem,
-	type ProviderModel,
-	type SaveProviderSettingsActionRequest,
-	type TextContent,
-	type VoiceInputSelection,
+import type {
+	AddProviderActionRequest,
+	ITelemetryService,
+	ProviderCapability,
+	ProviderCatalogResponse,
+	ProviderConfigField,
+	ProviderConfigFieldPrimitive,
+	ProviderListItem,
+	ProviderModel,
+	SaveProviderSettingsActionRequest,
+	VoiceInputSelection,
 } from "@cline/shared";
 import { createOAuthClientCallbacks } from "../../auth/client";
 import {
@@ -32,7 +23,6 @@ import {
 	getCachedClineRecommendedModels,
 	peekClineRecommendedModels,
 } from "../../services/llms/cline-recommended-models";
-import { resolveProviderConfig } from "../../services/llms/provider-defaults";
 import type {
 	ModelInfo,
 	ProviderClient,
@@ -50,14 +40,22 @@ import {
 	writeModelsFile,
 } from "./local-provider-registry";
 import {
+	buildMediaGenerationModelCatalog,
+	clearMediaGenerationSelections,
+	resolveActiveMediaGenerationSettings,
+} from "./media-generation-service";
+import {
 	fetchModelIdsFromSource,
 	resolveModelsSourceUrl,
 } from "./model-source";
+import {
+	CLINE_PASS_PROVIDER_ID,
+	resolveProviderModelMap,
+} from "./provider-model-map";
 
 export { ensureCustomProvidersLoaded } from "./local-provider-registry";
 
 const CLINE_PROVIDER_ID = "cline";
-const CLINE_PASS_PROVIDER_ID = "cline-pass";
 
 export interface ListLocalProvidersOptions {
 	isClinePassEnabled?: boolean;
@@ -99,24 +97,6 @@ export interface TranscribeConfiguredVoiceInputRequest {
 export interface CreateConfiguredStreamingTranscriptionSessionRequest {
 	expiresAfterSeconds?: number;
 	abortSignal?: AbortSignal;
-}
-
-export interface GenerateConfiguredMediaRequest {
-	mediaType: "image";
-	prompt: string;
-	abortSignal?: AbortSignal;
-}
-
-export interface GenerateConfiguredMediaResult {
-	content: Array<TextContent | MediaContent>;
-	usage?: AgentUsage;
-}
-
-export interface ResolvedMediaGenerationTarget {
-	mediaType: "image";
-	selection: MediaModelSelection;
-	providerConfig: ProviderConfig;
-	model: ModelInfo;
 }
 
 // --- Small pure helpers ---
@@ -171,138 +151,12 @@ export function isDedicatedTranscriptionModel(
 	return model.operation === "transcription";
 }
 
-/**
- * Whether a catalog model is image-generation eligible and has an executable
- * provider operation. Explicit image-generation models remain eligible when
- * external modality metadata is stale; mixed language models must advertise
- * text input and image output. The transport check deliberately fails closed
- * for providers without a matching image-capable operation.
- */
-export function isUsableImageGenerationModel(
-	providerId: string,
-	model: ModelInfo,
-): boolean {
-	const operation = model.operation ?? "language";
-	if (operation !== "language" && operation !== "image-generation") {
-		return false;
-	}
-	if (operation !== "image-generation" && !modelProducesImages(model)) {
-		return false;
-	}
-	const family = model.metadata?.family;
-	return LlmsModels.builtinProviderSupportsModelOperation({
-		providerId,
-		modelId: model.id,
-		operation: model.operation,
-		operationModes: model.operationModes,
-		modalities: model.modalities,
-		family: typeof family === "string" ? family : undefined,
-		capabilities: model.capabilities,
-	});
-}
-
 function toSortedProviderModels(
 	modelMap: Record<string, ModelInfo>,
 ): ProviderModel[] {
 	return Object.entries(modelMap)
 		.sort(([a], [b]) => a.localeCompare(b))
 		.map(([modelId, info]) => toProviderModel(modelId, info));
-}
-
-async function resolveProviderModelMap(
-	providerId: string,
-	config?: ProviderConfig,
-): Promise<Record<string, ModelInfo>> {
-	const registeredModels = await LlmsModels.getModelsForProvider(providerId);
-	const isClinePass = providerId === CLINE_PASS_PROVIDER_ID;
-	if (!config && !isClinePass) {
-		return registeredModels;
-	}
-
-	const resolved = await resolveProviderConfig(
-		providerId,
-		{
-			loadLatestOnInit: isClinePass,
-			loadPrivateOnAuth: true,
-			failOnError: false,
-		},
-		config,
-	);
-
-	if (providerId === "litellm" && resolved?.knownModels) {
-		return resolved.knownModels;
-	}
-	if (isClinePass && resolved?.knownModels) {
-		return resolved.knownModels;
-	}
-
-	return resolved?.knownModels
-		? {
-				...registeredModels,
-				...resolved.knownModels,
-			}
-		: registeredModels;
-}
-
-async function resolveMediaGenerationSelection(
-	manager: ProviderSettingsManager,
-	mediaType: MediaGenerationType,
-	selection: MediaModelSelection | undefined,
-): Promise<ResolvedMediaGenerationTarget | undefined> {
-	if (mediaType !== "image") return undefined;
-	if (!selection) return undefined;
-
-	const providerId = selection.providerId.trim();
-	const modelId = selection.modelId.trim();
-	if (!providerId || !modelId) return undefined;
-
-	try {
-		const state = manager.read();
-		if (!state.providers[providerId]) return undefined;
-		const config = manager.getProviderConfig(providerId, {
-			includeKnownModels: false,
-		});
-		if (!config) return undefined;
-
-		const modelMap = await resolveProviderModelMap(providerId, config);
-		const model = modelMap[modelId];
-		if (!model || !isUsableImageGenerationModel(providerId, model)) {
-			return undefined;
-		}
-
-		return {
-			mediaType,
-			selection: { providerId, modelId },
-			providerConfig: {
-				...config,
-				modelId,
-				modelInfo: model,
-				knownModels: {
-					...(config.knownModels ?? {}),
-					[modelId]: model,
-				},
-			},
-			model,
-		};
-	} catch {
-		return undefined;
-	}
-}
-
-/** Resolve a currently executable configured media target without mutating settings. */
-export async function resolveConfiguredMediaGenerationTarget(
-	manager: ProviderSettingsManager,
-	mediaType: MediaGenerationType,
-): Promise<ResolvedMediaGenerationTarget | undefined> {
-	try {
-		return await resolveMediaGenerationSelection(
-			manager,
-			mediaType,
-			manager.getMediaGenerationSettings()?.[mediaType],
-		);
-	} catch {
-		return undefined;
-	}
 }
 
 function uniqueTrimmed(values?: string[]): string[] {
@@ -511,17 +365,7 @@ function clearProviderModeSelections(
 		delete state.modes.voiceInput;
 		mutated = true;
 	}
-	const mediaGeneration = state.modes.mediaGeneration;
-	for (const mediaType of MEDIA_GENERATION_TYPES) {
-		if (mediaGeneration?.[mediaType]?.providerId !== providerId) {
-			continue;
-		}
-		delete mediaGeneration[mediaType];
-		if (Object.keys(mediaGeneration).length === 0) {
-			delete state.modes.mediaGeneration;
-		}
-		mutated = true;
-	}
+	mutated = clearMediaGenerationSelections(state, providerId) || mutated;
 	return mutated;
 }
 
@@ -951,21 +795,14 @@ export async function listLocalProviders(
 			(provider) => provider.id !== CLINE_PASS_PROVIDER_ID,
 		);
 	}
-	const imageModelsByProvider: Record<string, string[]> = {};
-	for (const provider of providers) {
-		const modelMap = providerEntries.find(
-			(entry) => entry.provider.id === provider.id,
-		)?.modelMap;
-		imageModelsByProvider[provider.id] = Object.entries(modelMap ?? {})
-			.filter(([, model]) => isUsableImageGenerationModel(provider.id, model))
-			.map(([modelId]) => modelId)
-			.sort((a, b) => a.localeCompare(b));
-	}
-	const mediaGenerationModels: MediaGenerationModelCatalog = {
-		audio: {},
-		image: imageModelsByProvider,
-		video: {},
-	};
+	const mediaGenerationModels = buildMediaGenerationModelCatalog(
+		providers.map((provider) => ({
+			providerId: provider.id,
+			models:
+				providerEntries.find((entry) => entry.provider.id === provider.id)
+					?.modelMap ?? {},
+		})),
+	);
 
 	const configuredVoiceInput = manager.getVoiceInputSettings();
 	const voiceProvider = configuredVoiceInput
@@ -981,21 +818,14 @@ export async function listLocalProviders(
 	);
 	const voiceInput =
 		configuredVoiceInput && voiceModel ? configuredVoiceInput : undefined;
-	const resolvedImageTarget = await resolveConfiguredMediaGenerationTarget(
+	const mediaGeneration = await resolveActiveMediaGenerationSettings(
 		manager,
-		"image",
+		new Set(
+			providers
+				.filter((provider) => provider.enabled)
+				.map((provider) => provider.id),
+		),
 	);
-	const imageProvider = resolvedImageTarget
-		? providers.find(
-				(provider) =>
-					provider.id === resolvedImageTarget.selection.providerId &&
-					provider.enabled,
-			)
-		: undefined;
-	const mediaGeneration =
-		resolvedImageTarget && imageProvider
-			? { image: resolvedImageTarget.selection }
-			: undefined;
 
 	return {
 		providers,
@@ -1098,91 +928,6 @@ export async function saveVoiceInputSettings(
 	const voiceInput = { providerId, modelId };
 	manager.setVoiceInputSettings(voiceInput);
 	return { settingsPath: manager.getFilePath(), voiceInput };
-}
-
-export async function saveMediaGenerationSettings(
-	manager: ProviderSettingsManager,
-	mediaType: MediaGenerationType,
-	selection: MediaModelSelection | undefined,
-): Promise<{
-	settingsPath: string;
-	mediaGeneration?: MediaGenerationSettings;
-}> {
-	if (mediaType !== "image") {
-		throw new Error(
-			`Media generation type "${mediaType}" is not supported yet`,
-		);
-	}
-
-	const current = { ...manager.getMediaGenerationSettings() };
-	if (!selection) {
-		delete current[mediaType];
-		const mediaGeneration =
-			Object.keys(current).length > 0 ? current : undefined;
-		manager.setMediaGenerationSettings(mediaGeneration);
-		return { settingsPath: manager.getFilePath(), mediaGeneration };
-	}
-
-	const providerId = selection.providerId.trim();
-	const modelId = selection.modelId.trim();
-	if (!providerId || !modelId) {
-		throw new Error("Image generation provider and model are required");
-	}
-	const target = await resolveMediaGenerationSelection(manager, "image", {
-		providerId,
-		modelId,
-	});
-	if (!target) {
-		throw new Error(
-			`Model "${modelId}" is not an executable image-generation model for provider "${providerId}"`,
-		);
-	}
-
-	const mediaGeneration = { ...current, [mediaType]: target.selection };
-	manager.setMediaGenerationSettings(mediaGeneration);
-	return { settingsPath: manager.getFilePath(), mediaGeneration };
-}
-
-/**
- * Generate media with the provider and model selected in server-side settings.
- *
- * The selection is revalidated immediately before each request so a removed,
- * disabled, or newly incompatible model cannot be invoked from stale session
- * state. Provider credentials never enter the tool input or result.
- */
-export async function generateConfiguredMedia(
-	manager: ProviderSettingsManager,
-	request: GenerateConfiguredMediaRequest,
-): Promise<GenerateConfiguredMediaResult> {
-	const target = await resolveConfiguredMediaGenerationTarget(
-		manager,
-		request.mediaType,
-	);
-	if (!target) {
-		throw new Error(
-			"The configured image generation provider or model is unavailable; choose one in Settings",
-		);
-	}
-
-	const result = await LlmsModels.generateMedia({
-		providerConfig: target.providerConfig,
-		modelId: target.selection.modelId,
-		prompt: request.prompt,
-		mediaType: request.mediaType,
-		abortSignal: request.abortSignal,
-	});
-
-	const mediaLabel = result.media.length === 1 ? "image" : "images";
-	return {
-		content: [
-			{
-				type: "text",
-				text: `Generated ${result.media.length} ${mediaLabel} with ${target.selection.providerId}/${target.selection.modelId}.`,
-			},
-			...result.media.map((media): MediaContent => ({ type: "media", media })),
-		],
-		...(result.usage ? { usage: result.usage } : {}),
-	};
 }
 
 export async function transcribeConfiguredVoiceInput(
